SC-246: Refactor SlideCollection tests
