+++ ShapeCrawler/Collections/IShapeCollection.cs
@@ -24,18 +24,18 @@ namespace ShapeCrawler
         /// <summary>
         ///     Create a new audio shape from stream and adds it to the end of the collection.
         /// </summary>
-        /// <param name="xPixel">The X coordinate for the left side of the shape.</param>
+        /// <param name="xPixels">The X coordinate for the left side of the shape.</param>
         /// <param name="yPixels">The Y coordinate for the left side of the shape.</param>
         /// <param name="mp3Stream">Audio stream data.</param>
-        IAudioShape AddNewAudio(int xPixel, int yPixels, Stream mp3Stream);
+        IAudioShape AddNewAudio(int xPixels, int yPixels, Stream mp3Stream);
 
         /// <summary>
         ///     Create a new video shape from stream and adds it to the end of the collection.
         /// </summary>
-        /// <param name="xPixel">The X coordinate for the left side of the shape.</param>
+        /// <param name="xPixels">The X coordinate for the left side of the shape.</param>
         /// <param name="yPixels">The Y coordinate for the left side of the shape.</param>
         /// <param name="videoStream">Video stream data.</param>
-        IVideoShape AddNewVideo(int xPixel, int yPixels, Stream videoStream);
+        IVideoShape AddNewVideo(int xPixels, int yPixels, Stream videoStream);
 
         /// <summary>
         ///     Get shape by identifier.
